crypto: api - Do not clear type bits in crypto_larval_lookup
Currently all bits not set in mask are cleared in crypto_larval_lookup.
This is unnecessary as wherever the type bits are used it is always
masked anyway.

This patch removes the clearing so that we may use bits set in the
type but not in the mask for special purposes, e.g., picking up
internal algorithms.

Change-Id: I0ae6fecdf937126901223b8aca489035ff2cf6b3
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>