crypto: api - Do not clear type bits in crypto_larval_lookup
Currently all bits not set in mask are cleared in crypto_larval_lookup.
This is unnecessary as wherever the type bits are used it is always
masked anyway.

This patch removes the clearing so that we may use bits set in the
type but not in the mask for special purposes, e.g., picking up
internal algorithms.

Change-Id: I3d705283af261ee9ce829aed2999f7c61489e999
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: kdrag0n <dragon@khronodragon.com>