clk: qcom: cpu-osm: Notify the current frequency to the topology driver
The topology driver keeps track of the current capacity of a CPU and
make it available via arch_scale_freq_capacity(). The PELT load
tracking in the scheduler and energy-aware task placement are the
clients of this API. Hence notify the new frequency to the topology
driver during frequency transition.

Change-Id: Ia0890956f59082496e7d5d7ba8a9c09b343e44ec
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>
Signed-off-by: Jason Edson <jaysonedson@gmail.com>