msm: kgsl: Increase worker thread priority
Currently, the kgsl worker thread is erroneosly ranked right below
Android's audio threads in terms of priority.

The kgsl worker thread is in the critical path for rendering frames to the
display, so increase its priority to match the priority of MDSS' kthread (mdss_fb0).

0ctobot:
Reduce priority to 6 in order to retain surfaceflinger and audio
thread preemption without needlessly interfering with more
important work.

Change-Id: I41f5d197bf8ebb3a5f3680ee54b0bebfd083f3e0
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>