f2fs/rapid_gc: do not wait until all invalid blocks are discarded
Undiscarded blocks may not reach 0 for various reasons, such as
discard granularity settings and discard thread terminating before GC ends,
and when that happens, GC thread enters an indefinite loop.

Remove this logic entirely.

This is not much of an issue since discard thread does not rely on rapid GC.

Change-Id: I7d2234b5dfbe9d3f2f5012b31f26b076299559ce
Signed-off-by: Park Ju Hyung <qkrwngud825@gmail.com>