drm/msm: Add a simple display state tracker for external use
Being able to easily call a function to query the current display state
can be useful for several reasons:
 - Removing the need a separate msm_drm_notifier in every driver that
 wants to query the display state
 - Reducing risk of widespread breakage if msm_drm_notifier's semantics
 change
 - Simplifying driver code
 - Reducing code duplication among drivers that need such functionality
 - Reducing code executed when the msm_drm_notifier chain is called in
 msm_drm as it is blocking and can stall wake

This is a reference implementation of the display_state interface for
msm_drm. Feel free to adapt it to other display stacks as needed. Note
that the high notifier priority (to prevent other drivers from breaking)
and use of the __read_mostly attribute (0.4% of operations are writes,
and this can be considered a hot path) are deliberate design choices.

Change-Id: I8d28d99e6d3fafde5389b05f40a4a50e6b983694
Signed-off-by: Danny Lin <danny@kdrag0n.dev>