cpu_input_boost: Allow configuration of each cluster's max-boost freq
On certain CPUs, the highest frequency could be very inefficient, and
thus frequently boosting to the frequency for max boosts can incur a
high power cost. To mitigate this, allow configuration of the max-boost
frequency for each CPU cluster, so that a more efficient frequency can
be selected for max boosts.

Change-Id: I0b791bacb46e4e8c2e01bf2d9f98e3fb76fa03e3
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Danny Lin <danny@kdrag0n.dev>