cpu_input_boost: Kill SchedTune boosting code
We'll rely on userspace's PowerHAL to perform SchedTune boosting
instead, since the default values of 10/50 are quite efficient in
practice and offer significant performance benefits. Userspace is also
more context-aware and thus can choose when to boost and unboost at more
appropriate times, depending on how the device is being used. For
example, it knows exactly when to unboost after starting a video,
whereas the kernel relies on a timeout to achieve this.

Change-Id: Ia146056f48c9b9fc0c78257d5d3b72817e5c9697
Signed-off-by: Danny Lin <danny@kdrag0n.dev>